making fixtures work
