+++ acceptance-tests/src/test/java/tech/pegasys/signers/keystorage/hashicorp/tests/HashicorpVaultAccessAcceptanceTest.java
@@ -14,15 +14,14 @@ package tech.pegasys.signers.keystorage.hashicorp.tests;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
+import tech.pegasys.signers.hashicorp.HashicorpConnection;
+import tech.pegasys.signers.hashicorp.HashicorpConnectionFactory;
+import tech.pegasys.signers.hashicorp.config.HashicorpKeyConfig;
+import tech.pegasys.signers.hashicorp.config.loader.toml.TomlConfigLoader;
 import tech.pegasys.signers.hashicorp.dsl.DockerClientFactory;
+import tech.pegasys.signers.hashicorp.dsl.HashicorpNode;
 import tech.pegasys.signers.hashicorp.dsl.certificates.CertificateHelpers;
-import tech.pegasys.signers.hashicorp.dsl.hashicorp.HashicorpNode;
 import tech.pegasys.signers.hashicorp.util.HashicorpConfigUtil;
-import tech.pegasys.signing.hashicorp.HashicorpConnection;
-import tech.pegasys.signing.hashicorp.HashicorpConnectionFactory;
-import tech.pegasys.signing.hashicorp.config.HashicorpKeyConfig;
-import tech.pegasys.signing.hashicorp.config.loader.toml.TomlConfigLoader;
-import tech.pegasys.signing.hashicorp.util.HashicorpConfigUtil;
 
 import java.io.IOException;
 import java.nio.file.Path;
