+++ keystorage/build.gradle
@@ -1,1 +1,1 @@
-//jar { enabled = false }
+jar { enabled = false }
