+++ keystorage/hashicorp/build.gradle
@@ -53,5 +53,4 @@ dependencies {
   testFixturesImplementation 'org.assertj:assertj-core'
   testFixturesImplementation 'org.awaitility:awaitility'
   testFixturesImplementation 'org.apache.tuweni:tuweni-net'
-  testFixturesImplementation 'org.mockito:mockito-core'
 }
