+++ keystorage/hashicorp/src/test/java/tech/pegasys/signers/hashicorp/TrustStoreTypeTest.java
@@ -1,0 +1,34 @@
+/*
+ * Copyright 2020 ConsenSys AG.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License"); you may not use this file except in compliance with
+ * the License. You may obtain a copy of the License at
+ *
+ * http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software distributed under the License is distributed on
+ * an "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied. See the License for the
+ * specific language governing permissions and limitations under the License.
+ */
+package tech.pegasys.signers.hashicorp;
+
+import static org.junit.jupiter.api.Assertions.assertEquals;
+
+import java.util.Optional;
+
+import org.junit.jupiter.api.Test;
+
+class TrustStoreTypeTest {
+
+  @Test
+  void trustStoreTypeCreatedFromValidValue() {
+    assertEquals(Optional.of(TrustStoreType.PEM), TrustStoreType.fromString("PEM"));
+  }
+
+  @Test
+  void turstStoreTypeEmptyFromInvalidValue() {
+    assertEquals(Optional.empty(), TrustStoreType.fromString("X"));
+    assertEquals(Optional.empty(), TrustStoreType.fromString("pem"));
+    assertEquals(Optional.empty(), TrustStoreType.fromString(null));
+  }
+}
