+++ keystorage/hashicorp/src/testFixtures/java/tech/pegasys/signers/hashicorp/dsl/HashicorpVaultTokens.java
@@ -1,0 +1,31 @@
+/*
+ * Copyright 2020 ConsenSys AG.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License"); you may not use this file except in compliance with
+ * the License. You may obtain a copy of the License at
+ *
+ * http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software distributed under the License is distributed on
+ * an "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied. See the License for the
+ * specific language governing permissions and limitations under the License.
+ */
+package tech.pegasys.signers.hashicorp.dsl;
+
+public class HashicorpVaultTokens {
+  private final String unsealKey;
+  private final String rootToken;
+
+  public HashicorpVaultTokens(final String unsealKey, final String rootToken) {
+    this.unsealKey = unsealKey;
+    this.rootToken = rootToken;
+  }
+
+  public String getUnsealKey() {
+    return unsealKey;
+  }
+
+  public String getRootToken() {
+    return rootToken;
+  }
+}
