+++ keystorage/hashicorp/src/testFixtures/java/tech/pegasys/signers/hashicorp/util/HashicorpConfigUtil.java
@@ -15,16 +15,16 @@ package tech.pegasys.signers.hashicorp.util;
 import static java.nio.charset.StandardCharsets.UTF_8;
 import static java.nio.file.Files.createTempFile;
 import static java.nio.file.Files.writeString;
-import static tech.pegasys.signing.hashicorp.config.loader.toml.TomlConfigLoader.PROP_HASHICORP_KEY_NAME;
-import static tech.pegasys.signing.hashicorp.config.loader.toml.TomlConfigLoader.PROP_HASHICORP_KEY_PATH;
-import static tech.pegasys.signing.hashicorp.config.loader.toml.TomlConfigLoader.PROP_HASHICORP_SERVER_HOST;
-import static tech.pegasys.signing.hashicorp.config.loader.toml.TomlConfigLoader.PROP_HASHICORP_SERVER_PORT;
-import static tech.pegasys.signing.hashicorp.config.loader.toml.TomlConfigLoader.PROP_HASHICORP_TIMEOUT;
-import static tech.pegasys.signing.hashicorp.config.loader.toml.TomlConfigLoader.PROP_HASHICORP_TLS_ENABLE;
-import static tech.pegasys.signing.hashicorp.config.loader.toml.TomlConfigLoader.PROP_HASHICORP_TLS_TS_PASSWORD;
-import static tech.pegasys.signing.hashicorp.config.loader.toml.TomlConfigLoader.PROP_HASHICORP_TLS_TS_PATH;
-import static tech.pegasys.signing.hashicorp.config.loader.toml.TomlConfigLoader.PROP_HASHICORP_TLS_TS_TYPE;
-import static tech.pegasys.signing.hashicorp.config.loader.toml.TomlConfigLoader.PROP_HASHICORP_TOKEN;
+import static tech.pegasys.signers.hashicorp.config.loader.toml.TomlConfigLoader.PROP_HASHICORP_KEY_NAME;
+import static tech.pegasys.signers.hashicorp.config.loader.toml.TomlConfigLoader.PROP_HASHICORP_KEY_PATH;
+import static tech.pegasys.signers.hashicorp.config.loader.toml.TomlConfigLoader.PROP_HASHICORP_SERVER_HOST;
+import static tech.pegasys.signers.hashicorp.config.loader.toml.TomlConfigLoader.PROP_HASHICORP_SERVER_PORT;
+import static tech.pegasys.signers.hashicorp.config.loader.toml.TomlConfigLoader.PROP_HASHICORP_TIMEOUT;
+import static tech.pegasys.signers.hashicorp.config.loader.toml.TomlConfigLoader.PROP_HASHICORP_TLS_ENABLE;
+import static tech.pegasys.signers.hashicorp.config.loader.toml.TomlConfigLoader.PROP_HASHICORP_TLS_TS_PASSWORD;
+import static tech.pegasys.signers.hashicorp.config.loader.toml.TomlConfigLoader.PROP_HASHICORP_TLS_TS_PATH;
+import static tech.pegasys.signers.hashicorp.config.loader.toml.TomlConfigLoader.PROP_HASHICORP_TLS_TS_TYPE;
+import static tech.pegasys.signers.hashicorp.config.loader.toml.TomlConfigLoader.PROP_HASHICORP_TOKEN;
 
 import java.io.IOException;
 import java.nio.file.Path;
